Fix #3, fix #11: key validation and improved cron.
- Issue #3: key validation logic was embed into the queue requests. Whenever
a NEW job run and the POST request to ccss.com's generate endpoint is
successful, the key is validated automatically. As such,
ao_ccss_validate_key() now just check for key status transient to
permform the internal validation. When the transient timeouts, the
validation process restart.
- Issue #11: as the maintenance cron job does a bit more then just truncate
log files, it was renamed accordingly.
- Update translations and stats.